Make all compounds functions
This changes the public interface but this is good:

This solves a naming issue (BatchPutItem was a function named like a
class to be consistent with BatchGetItemIterator. BatchGetItemIterator
was a class only because of an implementation detail)

Conceptually, it's always been clear that compounds receive the
connection as an argument and therefore were already *used as* functions.

Now they *are* functions.